OGM-310 Adapting test package names to reflect upgrade to WildFly
