+++ integrationtest/mongodb/src/test/java/org/hibernate/ogm/test/integration/wildfly/MongoDBModuleMemberRegistrationIT.java
@@ -18,8 +18,10 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
  * MA  02110-1301, USA.
  */
-package org.hibernate.ogm.test.integration.jbossas7;
+package org.hibernate.ogm.test.integration.wildfly;
 
+import org.hibernate.ogm.test.integration.wildfly.model.Member;
+import org.hibernate.ogm.test.integration.wildfly.util.ModuleMemberRegistrationDeployment;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.shrinkwrap.api.Archive;
@@ -28,11 +30,9 @@ import org.jboss.shrinkwrap.descriptor.api.persistence20.PersistenceDescriptor;
 import org.jboss.shrinkwrap.descriptor.api.persistence20.PersistenceUnit;
 import org.jboss.shrinkwrap.descriptor.api.persistence20.Properties;
 import org.junit.runner.RunWith;
-import org.hibernate.ogm.test.integration.jbossas7.model.Member;
-import org.hibernate.ogm.test.integration.jbossas7.util.ModuleMemberRegistrationDeployment;
 
 /**
- * Test the hibernate OGM module in JBoss AS using MongoDB
+ * Test for the Hibernate OGM module in WildFly using MongoDB
  *
  * @author Guillaume Scheibel <guillaume.scheibel@gmail.com>
  */
